api/requirements.txt
click==8.1.8
colorama==0.4.6
dlib==19.22.99
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==2.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.5.5.64
pillow==10.4.0
PyMySQL==1.0.2
setuptools==75.1.0
Werkzeug==2.1.1
wheel==0.44.0
zipp==3.20.2
#hashlib==3.4.0
#python==3.8.0